NEW METHOD: Pipes fixed
The old method of drawing 2 pipes was scrapped because I couldnt figure it out. But that aside, the new method draws one full rect = to the screen height, and then another rect ,with the same width but the height of the PYPE_GAP, drawn optop of it. The starting y-point of the gap rect is calculated and then drawn to the specifications provided.